Modify Registry::Find so that it can use views for faster searching
