+++ Sprocket/Scene/ECS.h
@@ -103,10 +103,6 @@ public:
     // Number of active entities in the registry.
     std::size_t Size() const;
 
-    // Returns the first entity satisfying the given predicate, or ECS::Null if
-    // none is found.
-    Entity Find(const EntityPredicate& pred);
-
     // Generates all active entities. This is fast, however adding and removing
     // entities while iterating results is undefined.
     cppcoro::generator<Entity> Fast();
@@ -121,6 +117,10 @@ public:
     // will be yielded. This should only be used for modifying the components, not
     // adding/removing new ones.
     template <typename Comp, typename... Rest> cppcoro::generator<Entity> View();
+
+    // Returns the first entity satisfying the given predicate, or ECS::Null if
+    // none is found. Can optionally provide components to filter on.
+    template <typename... Comps> Entity Find(const EntityPredicate& pred);
 
     template <typename Comp> void OnAdd(const EntityCallback& cb);
     template <typename Comp> void OnRemove(const EntityCallback& cb);
@@ -158,6 +158,26 @@ cppcoro::generator<Entity> Registry::View()
             co_yield entity;
         }
     }
+}
+
+template <typename... Comps>
+Entity Registry::Find(const EntityPredicate& pred)
+{
+    if constexpr (sizeof...(Comps) == 0) {
+        for (auto entity : Fast()) {
+            if (pred(entity)) {
+                return entity;
+            }
+        }
+    }
+    else {
+        for (auto entity : View<Comps...>()) {
+            if (pred(entity)) {
+                return entity;
+            }
+        }
+    }
+    return ECS::Null;
 }
 
 // ENTITY TEMPLATES
