--- Sprocket/Scene/SystemList.cpp
@@ -1,9 +1,0 @@
-#include "SystemList.h"
-
-namespace Sprocket {
-namespace ECS {
-
-
-
-}
-}
